Move API client to client.py
